Rename cilogon_ldap to ldap_data since I plan to move the LIGO LDAP stuff to that module
